Add comparability with Solarium v3.5
The class layout was refactored in Solarium v3.5, the
`Solarium\Core\Query\Query` was deprecated and replaced with a stub.
Classes of type `Solarium\QueryType\Select\Query\Query` are used
instead. This breaks tests for the type of `query`.  Since both of the
aforementioned classes implement `QueryInterface`, testing for instances
of that interface fixes the bug while maintaining backward
comparability.